projeto_PI/codigo_paginas_admin/requirements.txt
streamlit-authenticator
streamlit
pydantic
pydantic[email]
sqlalchemy
python-dotenv
pymysql 
pandas 
numpy